% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongoTable.R
\name{mongoTable}
\alias{mongoTable}
\title{mongoTable}
\usage{
mongoTable(
  connection,
  x,
  y = NULL,
  query = "{}",
  lowerize = FALSE,
  limit = NULL,
  sort = FALSE,
  decreasing = TRUE
)
}
\arguments{
\item{connection}{character. A mongo connection object initiated with mongolite::mongo().}

\item{x}{character. A field variable for which frequencies should be counted.}

\item{y}{character. An optional second field variable for which frequencies should be counted.}

\item{query}{character. An optional MongoDB query for data subset selection (e.g.: '\{\"year\": 2024\}').}

\item{lowerize}{logical. All levels in one dimensional tables will be lowerized.}

\item{limit}{integer. Defines the maximum length/dimensions of output.}

\item{sort}{logical. If TRUE, the output is sorted by frequency.}

\item{decreasing}{logical. If TRUE and sort==TRUE, the output is returned with decreasing frequencies. If TRUE and sort==FALSE, level names are returned in decreasing manner.}
}
\value{
A one or two dimensional frequency table.
}
\description{
Function to create one and two dimensional frequency tables from a MongoDB connection.
}
\examples{
# use mongolite::mongo() to connect to a MongoDB instance (demo server)
mon <- mongolite::mongo("mtcars", url =
"mongodb+srv://readwrite:test@cluster0-84vdt.mongodb.net/test")
if(mon$count() > 0) mon$drop()
mon$insert(mtcars)
stopifnot(mon$count() == nrow(mtcars))

############################################
## Create a one dimensional frequency table 
# for all x
mongoTable(connection = "mon", x = "cyl")

# create a one dimensional frequency table for all x matching a query
mongoTable(connection="mon", x="cyl", query = '{\"mpg\": {\"$gt": 20}}')

############################################
## Create a two dimensional frequency table

# for all x and y
mongoTable(con = "mon", x = "cyl", y = "gear")

# for all x and y matching a query
mongoTable(con="mon", x = "cyl", y = "gear", query = '{\"mpg\": {\"$gt": 20}}')
}
\keyword{MongoDB}
\keyword{frequency}
\keyword{table}
